Modified the TIA emulation so that it supports the PAL color loss effect.  Color loss occurs when the previous frame contains an odd number of scanlines.
The color loss effect is implemented using an idea from Eckhard
Stolberg: "I thought you could have each colour twice in the palette.
In the PAL palette the second colour would only be a shade of
grey. Then when something gets written to the TIA colour registers,
you could set bit0 of the colour according to bit0 of the
scanline counter for the last frame. That way an odd number of lines
would select the grey tones, while an even number would select the
normal colours."