Added basic message code to the TIA.  Can be called with MediaSource.showMessage(string& message uInt32 duration), where duration is the number of frames to display the message.
All messages are converted to upper case, since the fonts are very limited.
Fonts exist for A..Z and 0..9 only.  Some fonts are not the clearest, but its
the best I can do with only 4-pixel widths :)  Besides, they have an 'Atari'
look.

Updated the SDL and X11 versions to print messages under the following
conditions:

1) switching color modes (BW or color)
2) setting left and right difficulty for controllers
3) state loading/saving/changing (see below for this one)

Added methods that form the beginning of state loading and saving.  This
functionality is NOT currently present, and won't be for the 1.2 release.  The
methods added are for state saving, state slot changing, and state loading, and
are tied to function keys F9, F10, F11, respectively.  For now, they don't
actually do anything :)