Added the ability to change the aspect ratio of the screen in the SDL OpenGL port.  Normally, the window is double the native TIA frame width, so we always get a ratio of 2.0 (2:1).
Now, if you're using OpenGL mode, the '-gl_aspect' commandline argument
can take a decimal number representing the ratio you want to use.
For example, normal TV mode would be 1.3333 (4:3).

Depending on the resolution you use, you can make the window look much
more like the original Atari.  I find that in 1600x1200 mode, an
aspect of 1.6 looks quite authentic.

This is only supported in the SDL OpenGL port for now, since changing
the aspect took only 1 line of code (I love 3D graphics :)  It may
never be supported in software SDL, since it is quite a bit harder
to do it in software.

Ditto for the Cyberstella DirectDraw7 code (it will probably never be
implemented, unless hardware scaling is available).  If/when I get
around to the Direct3D port, it will be trivial to do there as well.