Updated the Cyberstella port wrt the recent Sound class changes.
Added code to draw menus to FrameBufferWin32.  This will be
changed around soon, when I get to implementing the windowed
modes.  I added it now to make sure that remapping works from
the in-game GUI (it does!).

Most of the core features are now working.  I need to add
windowed mode (as mentioned above), then start cleaning up
the settings and getting rid of the registry crap ...