Some further optimizations to the SDL OpenGL port.  I've reduced maximum CPU usage from 14% to approx. 10%.  While that may not seem like much, it's actually a 28% improvement.
The only way for it to get much better is to switch to paletted OpenGL
textures, but since they aren't supported in all versions of OpenGL,
I may not even bother.

For reference, the software SDL port uses a maximum of 5% CPU.  I don't
see the OpenGL port ever getting that low.