- Added an option to the makefile to specify the number of CPU's to use with 'make -jN' when compiling on SMP or distcc-based machines.  Defaults to 1 CPU, and most people won't need to touch it.
 - Cleaned up some sound related build options in the makefile.  Now
there is only one sound option (SDL, since the codebase is standardizing
on SDL), and the OSS and ALSA drivers are being discontinued.

 - Moved the settings that were specific to Linux SDL port to the emucore
Settings class, since they are now relevant to all SDL ports (including
Windows).

 - Reworked the sound selection code in mainSDL.  Now the '-sound'
commandline and ini-file argument is a boolean, representing whether
sound is enabled or disabled.