Added the ability to sort (in alphabetical order) by Name, Manufacturer or Rarity columns in the game listview.  The currently selected sort column is also saved, and then used when StellaX starts again.
The current width of columns is also saved, and reused when StellaX starts
again.

Still TODO is get the titlebar close button to 'hook' into the
MainDlg::Quit() method, so that it works the same way as pressing
the 'Exit' button or pressing 'Alt-F4'.  Right now, pressing the close
button in the titlebar bypasses the Quit() method, and AFAICT also
the freeing of any memory.  This bug seems to have been present in StellaX
since the beginning.