Yet more tweaks to the vLine, hLine and Scrollbar stuff.  It now looks (and works) as intended.
Converted many of the OpenGL functions back to similar code in the
corresponding software methods.  So OpenGL mode can now draw text
and bitmaps in GUI mode.  It just wasn't worth the effort to separate
the rendering so much, and all we lose by going back is accelerated
alpha-blending.  Since it's just eye candy anyway, I don't really care.

This is in anticipation of merging the software and OpenGL modes.  Once
that's complete, both FrameBuffer classes will be much smaller, and
we'll be able to switch between software and OpenGL modes while the
program is running.  This is one of the last 'features' that is
currently only accessible from the commandline.  My goal is that one
*shouldn't* have to use the commandline to change options, but the
ability will still be there if they want to.