Added register names to argument processing. This means you can set e.g. location 80 to the contents of the X register by saying "ram 80 x". This also means that there's a small conflict: if you're inputting the value 0x0a (decimal 10), you need to either say "0a" or "A". Otherwise the parser thinks you're talking about the "a" register instead of the hex number "a".
Added dereference operator to arg processing. This works the same way as
in C: you prefix a "*" to an argument. So "80" means a literal hex 80,
as before, but you can say "*80" to mean "the memory location pointed to
by location 80"... this also works with the registers. Example: if the X
register holds the value ff, you can say "*x" and it will be treated as
the value ff. Unlike C, multiple levels of dereference are not supported.
Also, the result of a dereference is always a byte for now.

Added "eval" command to debugger prompt. This evaluates one or more
arguments (which may hex values or a labels, and may have a * in front)
and prints the results in hex, binary, and decimal, along with the label
that has that value, if any.