Cleaned up the Debugger class by moving all CPU related stuff to CpuDebug.  This is similar to the last commit, which moved all RAM stuff to RamDebug.
Removed D6502, since it's functionality is provided by CpuDebug and is
no longer required.

The only user-visible change is that the CPU registers in CPU tab now
track changes.  It may seem like a lot of work to just accomplish that,
but the infrastructure will make it much easier to add similar
functionality to new tabs.  Besides, we needed to step back a little
and clean up the codebase.

Still TODO is turn TIADebug into a real DebuggerSystem object, and
finish the TIA tab.