More work on CheckListWidget.  This widget now sends signals when it need to be refilled or a line has been (un)checked.  Also, checked lines are now shown correctly (scrolling the list scrolls the checkboxes as well).
Added two different styles to CheckboxWidget.  The first is as before,
but the new one doesn't draw the surrounding box and fills with a
solid color instead of using 'X'.  Methods have been added to
set/unset drawing of the surrounding box, and change to doing a fill
vs using the 'X'.