Added a commandmenu, currently tied to the '\' key.  Eventually, this will also be tied to a joystick button, so common operations will be selectable by two buttons (one to enter the menu, another to select an item).
The commandmenu is a set of buttons for the most common operations
in Stella (specifically, those assigned to the F1 .. F12 keys).
Due to the way Stella is designed, pressing one of these buttons
immediately exits the commandmenu and proceeds with emulation.
Sorry if this causes anyone problems, but it's too difficult to
work around.

Still TODO is add joystick emulates mouse code (movement is done
and will be committed soon, still working on joystick button to mouse
button).